backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
python-multipart
pydantic-ai-slim[openai,mcp]>=0.0.14
